chore: rename server package
