+++ cmd/playground/main.go
@@ -14,8 +14,8 @@ import (
 	"github.com/x1unix/go-playground/internal/builder"
 	"github.com/x1unix/go-playground/internal/builder/storage"
 	"github.com/x1unix/go-playground/internal/config"
-	"github.com/x1unix/go-playground/internal/langserver"
-	"github.com/x1unix/go-playground/internal/langserver/webutil"
+	"github.com/x1unix/go-playground/internal/server"
+	"github.com/x1unix/go-playground/internal/server/webutil"
 	"github.com/x1unix/go-playground/pkg/goplay"
 	"github.com/x1unix/go-playground/pkg/util/cmdutil"
 	"github.com/x1unix/go-playground/pkg/util/osutil"
@@ -84,15 +84,15 @@ func start(goRoot string, logger *zap.Logger, cfg *config.Config) error {
 	// Initialize API endpoints
 	r := mux.NewRouter()
 	apiRouter := r.PathPrefix("/api").Subrouter()
-	svcCfg := langserver.ServiceConfig{Version: Version}
-	langserver.NewAPIv1Handler(svcCfg, playgroundClient, packages, buildSvc).
+	svcCfg := server.ServiceConfig{Version: Version}
+	server.NewAPIv1Handler(svcCfg, playgroundClient, packages, buildSvc).
 		Mount(apiRouter)
 
 	apiv2Router := apiRouter.PathPrefix("/v2").Subrouter()
-	langserver.NewAPIv2Handler(playgroundClient, buildSvc).Mount(apiv2Router)
+	server.NewAPIv2Handler(playgroundClient, buildSvc).Mount(apiv2Router)
 
 	// Web UI routes
-	tplVars := langserver.TemplateArguments{
+	tplVars := server.TemplateArguments{
 		GoogleTagID: cfg.Services.GoogleAnalyticsID,
 	}
 	if tplVars.GoogleTagID != "" {
@@ -104,8 +104,8 @@ func start(goRoot string, logger *zap.Logger, cfg *config.Config) error {
 	}
 
 	assetsDir := cfg.HTTP.AssetsDir
-	indexHandler := langserver.NewTemplateFileServer(zap.L(), filepath.Join(assetsDir, langserver.IndexFileName), tplVars)
-	spaHandler := langserver.NewSpaFileServer(assetsDir, tplVars)
+	indexHandler := server.NewTemplateFileServer(zap.L(), filepath.Join(assetsDir, server.IndexFileName), tplVars)
+	spaHandler := server.NewSpaFileServer(assetsDir, tplVars)
 	r.Path("/").
 		Handler(indexHandler)
 	r.Path("/snippet/{snippetID:[A-Za-z0-9_-]+}").
