+++ internal/server/handler_v1.go
@@ -1,4 +1,4 @@
-package langserver
+package server
 
 import (
 	"context"
