+++ internal/server/handler_v2.go
@@ -1,4 +1,4 @@
-package langserver
+package server
 
 import (
 	"context"
