+++ internal/server/resources/version.go.txt
