+++ internal/server/tplhandler.go
@@ -1,4 +1,4 @@
-package langserver
+package server
 
 import (
 	"bytes"
